requirements.txt
blis==0.7.9
boto3==1.26.130
botocore==1.29.130
catalogue==1.0.2
certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
cmake==3.26.3
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.7
dask==2023.5.1
data==0.4
decorator==5.1.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
filelock==3.12.0
fonttools==4.39.3
fr-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-2.3.0/fr_core_news_sm-2.3.0.tar.gz
fsspec==2023.5.0
funcsigs==1.0.2
future==0.18.3
gensim==4.3.1
huggingface-hub==0.17.3
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
joypy==0.2.6
kiwisolver==1.4.4
latex==0.7.0
lit==16.0.3
locket==1.0.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
murmurhash==1.0.9
networkx==3.1
nltk==3.8.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
ordered-set==4.1.0
packaging==23.1
pandas==2.0.1
partd==1.4.0
Pillow==9.5.0
plac==1.1.3
preshed==3.0.8
protobuf==4.24.3
PyLaTeX==1.4.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-nlp==0.4.1
pytorch-pretrained-bert==0.6.2
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.30.0
s3transfer==0.6.1
sacremoses==0.0.53
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
sentencepiece==0.1.99
shutilwhich==1.1.0
six==1.16.0
smart-open==6.3.0
spacy==2.3.9
srsly==1.0.6
sympy==1.11.1
tempdir==0.7.1
tensorboardX==2.6.2.2
thinc==7.4.6
threadpoolctl==3.1.0
tokenizers==0.10.3
toolz==0.12.0
torch==2.0.1
tqdm==4.65.0
transformers==4.12.5
triton==2.0.0
typing_extensions==4.5.0
tzdata==2023.3
Unidecode==1.3.7
urllib3==1.26.15
wasabi==0.10.1
wget==3.2
zipp==3.15.0
